+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `Diagnostics.energy_injection_rate` for estimation of energy injection (or
   dissipation) rate due to external velocities / forcing.
 
+- Add `Diagnostics.helicity_spectrum` for estimation of the helicity spectrum.
+
 - Coarse-grained vorticity: print warning if smoothing scale is too small (compared to Ewald
   splitting scale) to guarantee good accuracy.